Lock Twitter gem to version from Github
Author of Twitter gem began yanking older versions of the Twitter gem
from rubygems.org in hopes of grabbing the attention of developers to
upgrade to the latest gem version, which uses the latest version of the
Twitter API. The old Twitter API will be shut off in March of 2013. This
commit locks the gem to the older version based on a tag on Github until
the gem can be updated to the latest version.