Update functional search VCR cassette and fix failures
The cassette used a search for carmel instead of the red keep which returned 0 results.
Rerecorded with the correct search and actual results.
This revealed that publications have no route and a few other models needed route fixes in the helper.

Also, removed the now unused "boss_search_carmel" VCR cassette.